requirements.txt
numpy~=1.19.2
pandas~=1.3.3
flask~=1.1.2
pillow~=8.3.1
tensorflow~=2.6.0
gunicorn~=20.1.0